+++ eiam-20211201/src/Models/GetCredentialProviderResponseBody/credentialProvider/credentialProviderConfig/OAuthProviderConfig.php
@@ -11,7 +11,42 @@ class OAuthProviderConfig extends Model
     /**
      * @var string
      */
+    public $authorizationEndpoint;
+
+    /**
+     * @var string
+     */
+    public $authorizationFlow;
+
+    /**
+     * @var string
+     */
     public $clientId;
+
+    /**
+     * @var string
+     */
+    public $discoveryUrl;
+
+    /**
+     * @var string
+     */
+    public $issuer;
+
+    /**
+     * @var string
+     */
+    public $pkceChallengeMethod;
+
+    /**
+     * @var bool
+     */
+    public $pkceEnabled;
+
+    /**
+     * @var string
+     */
+    public $providerVendor;
 
     /**
      * @var string
@@ -21,10 +56,23 @@ class OAuthProviderConfig extends Model
     /**
      * @var string
      */
+    public $systemRedirectUri;
+
+    /**
+     * @var string
+     */
     public $tokenEndpoint;
     protected $_name = [
+        'authorizationEndpoint' => 'AuthorizationEndpoint',
+        'authorizationFlow' => 'AuthorizationFlow',
         'clientId' => 'ClientId',
+        'discoveryUrl' => 'DiscoveryUrl',
+        'issuer' => 'Issuer',
+        'pkceChallengeMethod' => 'PkceChallengeMethod',
+        'pkceEnabled' => 'PkceEnabled',
+        'providerVendor' => 'ProviderVendor',
         'scope' => 'Scope',
+        'systemRedirectUri' => 'SystemRedirectUri',
         'tokenEndpoint' => 'TokenEndpoint',
     ];
 
@@ -36,12 +84,44 @@ class OAuthProviderConfig extends Model
     public function toArray($noStream = false)
     {
         $res = [];
+        if (null !== $this->authorizationEndpoint) {
+            $res['AuthorizationEndpoint'] = $this->authorizationEndpoint;
+        }
+
+        if (null !== $this->authorizationFlow) {
+            $res['AuthorizationFlow'] = $this->authorizationFlow;
+        }
+
         if (null !== $this->clientId) {
             $res['ClientId'] = $this->clientId;
         }
 
+        if (null !== $this->discoveryUrl) {
+            $res['DiscoveryUrl'] = $this->discoveryUrl;
+        }
+
+        if (null !== $this->issuer) {
+            $res['Issuer'] = $this->issuer;
+        }
+
+        if (null !== $this->pkceChallengeMethod) {
+            $res['PkceChallengeMethod'] = $this->pkceChallengeMethod;
+        }
+
+        if (null !== $this->pkceEnabled) {
+            $res['PkceEnabled'] = $this->pkceEnabled;
+        }
+
+        if (null !== $this->providerVendor) {
+            $res['ProviderVendor'] = $this->providerVendor;
+        }
+
         if (null !== $this->scope) {
             $res['Scope'] = $this->scope;
+        }
+
+        if (null !== $this->systemRedirectUri) {
+            $res['SystemRedirectUri'] = $this->systemRedirectUri;
         }
 
         if (null !== $this->tokenEndpoint) {
@@ -59,12 +139,44 @@ class OAuthProviderConfig extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['AuthorizationEndpoint'])) {
+            $model->authorizationEndpoint = $map['AuthorizationEndpoint'];
+        }
+
+        if (isset($map['AuthorizationFlow'])) {
+            $model->authorizationFlow = $map['AuthorizationFlow'];
+        }
+
         if (isset($map['ClientId'])) {
             $model->clientId = $map['ClientId'];
         }
 
+        if (isset($map['DiscoveryUrl'])) {
+            $model->discoveryUrl = $map['DiscoveryUrl'];
+        }
+
+        if (isset($map['Issuer'])) {
+            $model->issuer = $map['Issuer'];
+        }
+
+        if (isset($map['PkceChallengeMethod'])) {
+            $model->pkceChallengeMethod = $map['PkceChallengeMethod'];
+        }
+
+        if (isset($map['PkceEnabled'])) {
+            $model->pkceEnabled = $map['PkceEnabled'];
+        }
+
+        if (isset($map['ProviderVendor'])) {
+            $model->providerVendor = $map['ProviderVendor'];
+        }
+
         if (isset($map['Scope'])) {
             $model->scope = $map['Scope'];
+        }
+
+        if (isset($map['SystemRedirectUri'])) {
+            $model->systemRedirectUri = $map['SystemRedirectUri'];
         }
 
         if (isset($map['TokenEndpoint'])) {