+++ eiam-20211201/src/Models/ListCloudAccountsResponseBody/cloudAccounts.php
@@ -6,6 +6,7 @@ namespace AlibabaCloud\SDK\Eiam\V20211201\Models\ListCloudAccountsResponseBody;
 
 use AlibabaCloud\Dara\Model;
 use AlibabaCloud\SDK\Eiam\V20211201\Models\ListCloudAccountsResponseBody\cloudAccounts\cloudAccountHealthCheckResult;
+use AlibabaCloud\SDK\Eiam\V20211201\Models\ListCloudAccountsResponseBody\cloudAccounts\privilegeHostingError;
 
 class cloudAccounts extends Model
 {
@@ -65,6 +66,26 @@ class cloudAccounts extends Model
     public $instanceId;
 
     /**
+     * @var string[]
+     */
+    public $privilegeApplicationIds;
+
+    /**
+     * @var privilegeHostingError
+     */
+    public $privilegeHostingError;
+
+    /**
+     * @var string
+     */
+    public $privilegeHostingState;
+
+    /**
+     * @var string
+     */
+    public $privilegeStatus;
+
+    /**
      * @var int
      */
     public $updateTime;
@@ -80,6 +101,10 @@ class cloudAccounts extends Model
         'createTime' => 'CreateTime',
         'description' => 'Description',
         'instanceId' => 'InstanceId',
+        'privilegeApplicationIds' => 'PrivilegeApplicationIds',
+        'privilegeHostingError' => 'PrivilegeHostingError',
+        'privilegeHostingState' => 'PrivilegeHostingState',
+        'privilegeStatus' => 'PrivilegeStatus',
         'updateTime' => 'UpdateTime',
     ];
 
@@ -87,6 +112,12 @@ class cloudAccounts extends Model
     {
         if (null !== $this->cloudAccountHealthCheckResult) {
             $this->cloudAccountHealthCheckResult->validate();
+        }
+        if (\is_array($this->privilegeApplicationIds)) {
+            Model::validateArray($this->privilegeApplicationIds);
+        }
+        if (null !== $this->privilegeHostingError) {
+            $this->privilegeHostingError->validate();
         }
         parent::validate();
     }
@@ -136,6 +167,29 @@ class cloudAccounts extends Model
 
         if (null !== $this->instanceId) {
             $res['InstanceId'] = $this->instanceId;
+        }
+
+        if (null !== $this->privilegeApplicationIds) {
+            if (\is_array($this->privilegeApplicationIds)) {
+                $res['PrivilegeApplicationIds'] = [];
+                $n1 = 0;
+                foreach ($this->privilegeApplicationIds as $item1) {
+                    $res['PrivilegeApplicationIds'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->privilegeHostingError) {
+            $res['PrivilegeHostingError'] = null !== $this->privilegeHostingError ? $this->privilegeHostingError->toArray($noStream) : $this->privilegeHostingError;
+        }
+
+        if (null !== $this->privilegeHostingState) {
+            $res['PrivilegeHostingState'] = $this->privilegeHostingState;
+        }
+
+        if (null !== $this->privilegeStatus) {
+            $res['PrivilegeStatus'] = $this->privilegeStatus;
         }
 
         if (null !== $this->updateTime) {
@@ -195,6 +249,29 @@ class cloudAccounts extends Model
 
         if (isset($map['InstanceId'])) {
             $model->instanceId = $map['InstanceId'];
+        }
+
+        if (isset($map['PrivilegeApplicationIds'])) {
+            if (!empty($map['PrivilegeApplicationIds'])) {
+                $model->privilegeApplicationIds = [];
+                $n1 = 0;
+                foreach ($map['PrivilegeApplicationIds'] as $item1) {
+                    $model->privilegeApplicationIds[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['PrivilegeHostingError'])) {
+            $model->privilegeHostingError = privilegeHostingError::fromMap($map['PrivilegeHostingError']);
+        }
+
+        if (isset($map['PrivilegeHostingState'])) {
+            $model->privilegeHostingState = $map['PrivilegeHostingState'];
+        }
+
+        if (isset($map['PrivilegeStatus'])) {
+            $model->privilegeStatus = $map['PrivilegeStatus'];
         }
 
         if (isset($map['UpdateTime'])) {